Move some code to utils
