+++ tested/languages/config.py
@@ -18,12 +18,8 @@ There are a few callbacks that must be implemented. These raise a
 `NotImplementedError`, so proper editors will warn you (or TESTed will crash when
 using your language).
 """
-import html
 import json
 import logging
-import math
-import os
-import re
 import sys
 import typing
 from collections import defaultdict
@@ -32,7 +28,7 @@ from pathlib import Path
 from typing import Callable, Dict, List, Mapping, Optional, Set, Tuple, Union
 
 from tested.datatypes import AdvancedTypes, AllTypes, ExpressionTypes, string_to_type
-from tested.dodona import AnnotateCode, ExtendedMessage, Message, Permission, Status
+from tested.dodona import AnnotateCode, Message, Status
 from tested.features import Construct
 from tested.internationalization import get_i18n_string
 from tested.languages.conventionalize import (
@@ -43,6 +39,7 @@ from tested.languages.conventionalize import (
     submission_name,
 )
 from tested.languages.description_generator import DescriptionGenerator
+from tested.languages.utils import limit_output, trace_to_html
 from tested.serialisation import ExceptionValue, FunctionCall, Statement, Value
 from tested.utils import fallback, get_args
 
@@ -55,102 +52,6 @@ FileFilter = Callable[[Path], bool]
 CallbackResult = Tuple[Command, Union[List[str], FileFilter]]
 
 _logger = logging.getLogger(__name__)
-
-
-def executable_name(basename: str) -> str:
-    """
-    Utility function that will
-
-    :param basename: The name of the executable without extension.
-
-    :return: The executable with extension corresponding to the platform.
-    """
-    if os.name == "nt":
-        return f"{basename}.exe"
-    else:
-        return basename
-
-
-def limit_output(
-    output: str,
-    limit_characters: int = 512,
-    max_lines: int = 20,
-    ellipsis_str: str = "...",
-) -> str:
-    """
-    Utility function for limiting a string output
-
-    :param output: String that possible needs to be abbreviated
-    :param limit_characters: Maximum characters used in the output
-    :param max_lines: Maximum lines in the output
-    :param ellipsis_str: ellipsis used when abbreviated is needed
-
-    :return: The abbreviated 'output' if needed otherwise the 'output' itself
-    """
-    lines = output.splitlines()
-    # Case character limit not exceeded and line limit not exceeded
-    if len(output) <= limit_characters and len(lines) <= max_lines:
-        return output
-    # Case character limit exceeded
-    max_chars = limit_characters - len(ellipsis_str)
-    forward_buffer = []
-    backward_buffer = []
-    len_lines = len(lines)
-    for f in range(math.ceil(min(max_lines - 1, len_lines) / 2)):
-        r = len_lines - f - 1
-        # Case last lines to consider are the same
-        if f == r:
-            forward_buffer.append(lines[f][: (max_chars - 1)])
-        # Otherwise
-        else:
-            next_line, prev_line = lines[f], lines[r]
-            current_length = len(next_line) + len(prev_line) + 2
-            # Both lines can be add in full
-            if current_length < max_chars:
-                forward_buffer.append(next_line)
-                backward_buffer.append(prev_line)
-                max_chars -= current_length
-            # Lines must be limited
-            else:
-                half = max_chars / 2
-                # Next line can be add in full
-                if len(next_line) + 2 < max_chars:
-                    forward_buffer.append(next_line)
-                    max_chars -= len(next_line) + 2
-                    backward_buffer.append(prev_line[-max_chars:])
-                # Prev line can be add in full
-                elif len(prev_line) + 2 < max_chars:
-                    backward_buffer.append(prev_line)
-                    max_chars -= len(prev_line) + 2
-                    forward_buffer.append(next_line[:max_chars])
-                # Both lines needed abbreviation
-                else:
-                    forward_buffer.append(next_line[: math.ceil(half - 1)])
-                    backward_buffer.append(prev_line[-math.floor(half - 1) :])
-                # Terminate loop because character limit reached
-                break
-    # Concat buffer
-    return "\n".join(forward_buffer + [ellipsis_str] + backward_buffer[::-1])
-
-
-def trace_to_html(
-    traceback: str,
-    link_regex: str = r"&lt;code&gt;:([0-9]+)",
-    link_subs: str = r'<a href="#" class="tab-link" data-tab="code" '
-    r'data-line="\1">&lt;code&gt;:\1</a>',
-) -> ExtendedMessage:
-    # Escape special characters
-    traceback = html.escape(traceback)
-    # Compile regex
-    link_regex = re.compile(link_regex)
-    # Add links to
-    traceback = link_regex.sub(link_subs, traceback)
-    _logger.debug(f"<pre><code>{traceback}</code></pre>")
-    return ExtendedMessage(
-        description=f"<pre><code>{traceback}</code></pre>",
-        format="html",
-        permission=Permission.STUDENT,
-    )
 
 
 class TypeSupport(Enum):
