+++ tested/languages/java/config.py
@@ -3,13 +3,18 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Dict, List, Tuple
 
 from tested.dodona import AnnotateCode, Message
-from tested.languages.config import CallbackResult, Command, Language, limit_output
+from tested.languages.config import CallbackResult, Command, Language
 from tested.languages.conventionalize import (
     Conventionable,
     NamingConventions,
     submission_file,
 )
-from tested.languages.utils import jvm_cleanup_stacktrace, jvm_memory_limit, jvm_stderr
+from tested.languages.utils import (
+    jvm_cleanup_stacktrace,
+    jvm_memory_limit,
+    jvm_stderr,
+    limit_output,
+)
 from tested.serialisation import FunctionCall, Statement, Value
 
 if TYPE_CHECKING:
